add new Command Tracker and organize mesage class locations
