+++ src/ReactiveDomain.Messaging/Command.cs
@@ -60,37 +60,4 @@ namespace ReactiveDomain.Messaging {
             return new Canceled(MsgId, GetType().FullName,Guid.Empty, CorrelationId, new SourceId(this));
         }
     }
-
-    /// <summary>
-    /// Indicates receipt of a command message at a command handler with a registered target.
-    /// Does not indicate success or failure of command processing.
-    /// </summary>
-    /// <inheritdoc cref="Message"/>
-    public class AckCommand : Message {
-        /// <summary>
-        /// MsgId of the Command being acked
-        /// </summary>
-        public readonly Guid CommandId;
-        /// <summary>
-        /// Full Type Name of the Command being acked
-        /// </summary>
-        public readonly string CommandFullName;
-        /// <summary>
-        /// Id of the Command Handler sending the ack
-        /// </summary>
-        public readonly Guid HandlerId;
-
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="commandId">MsgId of the Command being acked</param>
-        /// <param name="commandFullName">Full Type Name of the Command being acked</param>
-        /// <param name="handlerId">Id of the Command Handler sending the ack</param>
-        public AckCommand(Guid commandId, string commandFullName, Guid handlerId) {
-            CommandId = commandId;
-            CommandFullName = commandFullName;
-            HandlerId = handlerId;
-        }
-
-    }
 }
